tacd: move daemon code to separate file
